requirements.txt
PyPDF2>=3.0.0
Pillow>=9.0.0
pandas>=1.3.0
openpyxl>=3.0.0
xlsxwriter>=3.0.0
tqdm>=4.65.0
img2pdf>=0.4.0 